fix: BulkStockCountResponseDto stockCount Integer → Long 변경

+++ hub-service/src/main/java/com/boxoffice/hubservice/client/BulkStockCountResponseDto.java
@@ -2,4 +2,4 @@ package com.boxoffice.hubservice.client;
 
 import java.util.UUID;
 
-public record BulkStockCountResponseDto(UUID hubId, Integer stockCount) { }
+public record BulkStockCountResponseDto(UUID hubId, Long stockCount) { }
